backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
django-filter==24.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-dotenv==1.0.1
pytest==8.2.2
pytest-django==4.8.0
whitenoise==6.6.0
django-ratelimit==4.1.0
pytest-cov==4.1.0
channels==4.0.0
channels-redis==4.2.0
django-redis==5.4.0
daphne==4.1.0
django-jazzmin==3.0.0
mutagen==1.47.0
agora-token-builder==1.0.0
